requirements.txt
requests
mysql-connector-python
python-dotenv
pandas
streamlit
matplotlib
apscheduler
openpyxl